Refractoring Ongoing
