+++ API/AbokiAPI/Services/UserRepository.cs
@@ -1,0 +1,450 @@
+﻿using AbokiCore;
+using AbokiData.Configuration;
+using AbokiData.Models.DTOs.Requests;
+using AbokiData.Models.DTOs.Responses;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using Microsoft.IdentityModel.Tokens;
+using System;
+using System.Collections.Generic;
+using System.IdentityModel.Tokens.Jwt;
+using System.Linq;
+using System.Security.Claims;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AbokiAPI.Services
+{
+    public class UserRepository : IUserRepository
+    {
+        private readonly UserManager<ApplicationUser> _userManager;
+        private readonly JwtConfig _jwtConfig;
+        private readonly TokenValidationParameters _tokenValidationParams;
+        private readonly ApiDbContext _apiDbContext;
+        private readonly RoleManager<IdentityRole> _roleManager;
+        private readonly ILogger<UserRepository> _logger;
+
+        public UserRepository(
+           UserManager<ApplicationUser> userManager,
+           IOptionsMonitor<JwtConfig> optionsMonitor,
+           TokenValidationParameters tokenValidationParams,
+           ApiDbContext apiDbContext,
+            RoleManager<IdentityRole> roleManager,
+             ILogger<UserRepository> logger)
+        {
+            _userManager = userManager;
+            _jwtConfig = optionsMonitor.CurrentValue;
+            _tokenValidationParams = tokenValidationParams;
+            _apiDbContext = apiDbContext;
+            _roleManager = roleManager;
+            _logger = logger;
+        }
+
+        public async Task<RegistrationResponse> RegisterUserAsync(UserRegistrationDto user)
+        {
+
+            if (user == null)
+            {
+                throw new ArgumentNullException("user credital is empty");
+            }
+
+            if (user.Email != null)
+            {
+                //If email has been used
+                var existingUser = await _userManager.FindByEmailAsync(user.Email);
+
+                //if email exist, return bad request
+                if (existingUser != null)
+                {
+                    return new RegistrationResponse()
+                    {
+                        Errors = new List<string>()
+                         {
+                           "Email already in use"
+                         },
+                        Success = false
+                    };
+                }
+
+                //accept the data from the user
+                var newUser = new ApplicationUser()
+                {
+                    Email = user.Email,
+                    UserName = user.Username
+                };
+
+                //Create the user using CreateAsync Method
+                var isCreated = await _userManager.CreateAsync(newUser, user.Password);
+
+                if (isCreated.Succeeded)
+                {
+                    //add te user to a role
+                    await _userManager.AddToRoleAsync(newUser, "AppUser");
+
+
+                    var jwtToken = await GenerateJwtToken(newUser);
+
+                    return jwtToken;
+
+                }
+                else
+                {
+                    return new RegistrationResponse()
+                    {
+                        Errors = isCreated.Errors.Select(x => x.Description).ToList(),
+                        Success = false
+
+                    };
+                }
+
+
+            }
+
+            return new RegistrationResponse()
+            {
+                Errors = new List<string>()
+                {
+                    "Invalid Payload"
+                },
+                Success = false
+            };
+        }
+
+        public async Task<RegistrationResponse> LoginAsync(UserLoginRequest user)
+        {
+            if (user != null)
+            {
+                var existingUser = await _userManager.FindByEmailAsync(user.Email);
+
+                if (existingUser == null)
+                {
+                    return (new RegistrationResponse()
+                    {
+                        Errors = new List<string>()
+                       {
+                         "Invalid login request"
+                        },
+                        Success = false
+                    });
+                }
+
+                var isCorrect = await _userManager.CheckPasswordAsync
+                    (existingUser, user.Password);
+
+                if (!isCorrect)
+                {
+                    return (new RegistrationResponse()
+                    {
+                        Errors = new List<string>()
+                        {
+                            "Invalid login request"
+                        },
+                        Success = false
+                    });
+                }
+
+                var jwtToken = await GenerateJwtToken(existingUser);
+
+                return jwtToken;
+            }
+            return (new RegistrationResponse()
+            {
+                Errors = new List<string>()
+                {
+                    "Invalid payload"
+                },
+                Success = false
+            });
+        }
+
+        public async Task<AuthResult> RefreshTokenAsync(TokenRequest tokenRequest)
+        {
+            if (tokenRequest != null)
+            {
+                var result = await VerifyAndGenerateToken(tokenRequest);
+
+                if (result == null)
+                {
+                    return (new AuthResult()
+                    {
+                        Errors = new List<string>() {
+                            "Invalid tokens"
+                        },
+                        Success = false
+                    });
+                }
+
+                if (result.Errors != null && result.Errors.Count > 0)
+                {
+                    return result;
+                }
+
+                return (result);
+            }
+
+            return (new AuthResult()
+            {
+                Errors = new List<string>() {
+                    "Invalid payload"
+                },
+                Success = false
+            });
+        }
+
+
+
+
+
+
+
+
+
+        //Generate JWT Token
+        private async Task<RegistrationResponse> GenerateJwtToken(ApplicationUser user)
+        {
+            //instances of JwtTokenHandler
+            var jwtTokenHandler = new JwtSecurityTokenHandler();
+
+            //My_Key
+            var Key = Encoding.ASCII.GetBytes(_jwtConfig.Secret);
+
+            //Inject the claims to the user
+            var claims = await GetAllValidClaims(user);
+
+            var tokenDescriptor = new SecurityTokenDescriptor
+            {
+                Subject = new ClaimsIdentity(claims),
+                Expires = DateTime.UtcNow.AddSeconds(30),
+                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(Key), SecurityAlgorithms.HmacSha256Signature)
+            };
+
+            var token = jwtTokenHandler.CreateToken(tokenDescriptor);
+
+            var jwtToken = jwtTokenHandler.WriteToken(token);
+
+            var refreshToken = new RefreshToken()
+            {
+                JwtId = token.Id,
+                IsUsed = false,
+                IsRevorked = false,
+                UserId = user.Id,
+                AddedDate = DateTime.UtcNow,
+                ExpiryDate = DateTime.UtcNow.AddMonths(6),
+                Token = RandomString(35) + Guid.NewGuid()
+
+            };
+
+            await _apiDbContext.RefreshTokens.AddAsync(refreshToken);
+            await _apiDbContext.SaveChangesAsync();
+
+            return new RegistrationResponse()
+            {
+                Token = jwtToken,
+                Success = true,
+                RefreshToken = refreshToken.Token
+            };
+        }
+
+        //Get all valid claims for the corresponding user
+        private async Task<List<Claim>> GetAllValidClaims(ApplicationUser user)
+        {
+            //var _options = new IdentityOptions();
+
+            var claims = new List<Claim>
+            {
+                    new Claim("Id", user.Id),
+                    new Claim(JwtRegisteredClaimNames.Email, user.Email),
+                    new Claim(JwtRegisteredClaimNames.Sub, user.Email),
+                    new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString())
+            };
+
+            //Getting the claims that we have assigned to the user 
+            var userClaims = await _userManager.GetClaimsAsync(user);
+            claims.AddRange(userClaims);
+
+            //Get the user and add it to the claims
+
+            var userRoles = await _userManager.GetRolesAsync(user);
+
+            foreach (var userRole in userRoles)
+            {
+
+                var role = await _roleManager.FindByNameAsync(userRole);
+
+                if (role != null)
+                {
+                    claims.Add(new Claim(ClaimTypes.Role, userRole));
+
+                    var roleClaims = await _roleManager.GetClaimsAsync(role);
+                    foreach (var roleClaim in roleClaims)
+                    {
+                        claims.Add(roleClaim);
+                    }
+                }
+            }
+
+            return claims;
+
+        }
+
+        private async Task<AuthResult> VerifyAndGenerateToken(TokenRequest tokenRequest)
+        {
+            var jwtTokenHandler = new JwtSecurityTokenHandler();
+
+            try
+            {
+                // Validation 1 - Validation JWT token format
+                _tokenValidationParams.ValidateLifetime = false;
+                var tokenInVerification = jwtTokenHandler.ValidateToken(tokenRequest.Token, _tokenValidationParams, out var validatedToken);
+                _tokenValidationParams.ValidateLifetime = true;
+
+                // Validation 2 - Validate encryption alg
+                if (validatedToken is JwtSecurityToken jwtSecurityToken)
+                {
+                    var result = jwtSecurityToken.Header.Alg.Equals(SecurityAlgorithms.HmacSha256, StringComparison.InvariantCultureIgnoreCase);
+
+                    if (result == false)
+                    {
+                        return null;
+                    }
+                }
+
+                // Validation 3 - validate expiry date
+                var utcExpiryDate = long.Parse(tokenInVerification.Claims.FirstOrDefault(x => x.Type == JwtRegisteredClaimNames.Exp).Value);
+
+                var expiryDate = UnixTimeStampToDateTime(utcExpiryDate);
+
+                if (expiryDate > DateTime.UtcNow)
+                {
+                    return new AuthResult()
+                    {
+                        Success = false,
+                        Errors = new List<string>() {
+                            "Token has not yet expired"
+                        }
+                    };
+                }
+
+                // validation 4 - validate existence of the token
+                var storedToken = await _apiDbContext.RefreshTokens.FirstOrDefaultAsync(x => x.Token == tokenRequest.RefreshToken);
+
+                if (storedToken == null)
+                {
+                    return new AuthResult()
+                    {
+                        Success = false,
+                        Errors = new List<string>() {
+                            "Token does not exist"
+                        }
+                    };
+                }
+
+                // Validation 5 - validate if used
+                if (storedToken.IsUsed)
+                {
+                    return new AuthResult()
+                    {
+                        Success = false,
+                        Errors = new List<string>() {
+                            "Token has been used"
+                        }
+                    };
+                }
+
+                // Validation 6 - validate if revoked
+                if (storedToken.IsRevorked)
+                {
+                    return new AuthResult()
+                    {
+                        Success = false,
+                        Errors = new List<string>() {
+                            "Token has been revoked"
+                        }
+                    };
+                }
+
+                // Validation 7 - validate the id
+                var jti = tokenInVerification.Claims.FirstOrDefault(x => x.Type == JwtRegisteredClaimNames.Jti).Value;
+
+                if (storedToken.JwtId != jti)
+                {
+                    return new AuthResult()
+                    {
+                        Success = false,
+                        Errors = new List<string>() {
+                            "Token doesn't match"
+                        }
+                    };
+                }
+
+                // Validation 8 - validate stored token expiry date
+                if (storedToken.ExpiryDate < DateTime.UtcNow)
+                {
+                    return new AuthResult()
+                    {
+                        Success = false,
+                        Errors = new List<string>() {
+                            "Refresh token has expired"
+                        }
+                    };
+                }
+
+                // update current token 
+
+                storedToken.IsUsed = true;
+                _apiDbContext.RefreshTokens.Update(storedToken);
+                await _apiDbContext.SaveChangesAsync();
+
+                // Generate a new token
+                var dbUser = await _userManager.FindByIdAsync(storedToken.UserId);
+                return await GenerateJwtToken(dbUser);
+            }
+            catch (Exception ex)
+            {
+                if (ex.Message.Contains("Lifetime validation failed. The token is expired."))
+                {
+
+                    return new AuthResult()
+                    {
+                        Success = false,
+                        Errors = new List<string>() {
+                            "Token has expired please re-login"
+                        }
+                    };
+
+                }
+                else
+                {
+                    return new AuthResult()
+                    {
+                        Success = false,
+                        Errors = new List<string>() {
+                            "Something went wrong."
+                        }
+                    };
+                }
+            }
+        }
+
+
+        private DateTime UnixTimeStampToDateTime(long unixTimeStamp)
+        {
+            var dateTimeVal = new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Utc);
+            dateTimeVal = dateTimeVal.AddSeconds(unixTimeStamp).ToUniversalTime();
+
+            return dateTimeVal;
+        }
+
+
+        private string RandomString(int length)
+        {
+            var random = new Random();
+            var chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+            return new string(Enumerable.Repeat(chars, length)
+                .Select(x => x[random.Next(x.Length)]).ToArray());
+        }
+
+    }
+}
